+++ gtmfe/include/shapereader.h
@@ -1,14 +1,14 @@
 #ifndef _SHAPEREADER_H_
 #define _SHAPEREADER_H_
 
-extern double* SHAPEarray;
+extern float* SHAPEarray;
 extern int* SHAPEenergies;
 
 void readSHAPEarray(const char* filename, int seqlength);
 void free_shapeArray(int len);
 void print_shapeArray(int len);
-int calcShapeEnergy(double shapeNumber);
-double shapeModel(double SHAPE_value);
+int calcShapeEnergy(float shapeNumber);
+float shapeModel(float SHAPE_value);
 
 #ifdef __cplusplus
 extern "C"{
